Fix UX  "Video cropping has issues with Android 11's Gestures" see: https://bugs.telegram.org/c/1557
When Android 11's Gestures are enabled, trying to crop a landscape video in portrait mode is near impossible because the cropping handles are too close the edge. This means if you try to move the handles, you will end up doing the "back" gesture.
Also, you can't uncrop back to the original size because you can't drag the handles all the way to the sides.
Steps to reproduce
    1. Get a phone on Android 11, turn on Gestures
    2. Open a landscape video in portrait mode in the full screen media editor
    3. Tap the crop button
    4. Try to crop the video using the handles
Comment:
    It isn’t a bug but a user experience issue. There are two problem to fix it:
        1. Android introduced gesture control just at Android Q and above
        2. By the gesture API you can block default gestures only on areas with height 200dp(in sum).

    To see the changes you should try to swipe only in center of the crop rect edge(+-45dp vertically) area or at angles(handles)  - it’s Api restriction not my fault ¯\_(ツ)_/¯